refactor: divide util functions and components for stories
