+++ src/components/index.jsx
@@ -1,0 +1,3 @@
+export { GigPanel } from './GigPanel'
+export { LocationTitleSlide } from './LocationTitleSlide'
+export { Carousel } from './Carousel'
