+++ src/utils/getMelbourneDate.jsx
@@ -1,0 +1,14 @@
+function getMelbourneDate() {
+  return new Date()
+    .toLocaleDateString('en-AU', {
+      timeZone: 'Australia/Melbourne',
+      year: 'numeric',
+      month: '2-digit',
+      day: '2-digit',
+    })
+    .split('/')
+    .reverse()
+    .join('-')
+}
+
+export { getMelbourneDate }
