+++ src/utils/getPublicUrl.jsx
@@ -1,0 +1,7 @@
+import { LML_URL_TODAY } from '../constants/common'
+
+function getPublicUrl() {
+  return LML_URL_TODAY // Always use lml.live for the website URL
+}
+
+export { getPublicUrl }
