+++ src/utils/getSuburb.jsx
@@ -1,0 +1,6 @@
+function getSuburb(address) {
+  const match = address.match(/(?:,\s*)?([A-Za-z\s]+)(?:\s+\d{4})?$/)
+  return match ? match[1].trim() : ''
+}
+
+export { getSuburb }
