+++ src/utils/toTitleCase.jsx
@@ -1,0 +1,8 @@
+function toTitleCase(str) {
+  return str.replace(
+    /\w\S*/g,
+    (txt) => txt.charAt(0).toUpperCase() + txt.substr(1).toLowerCase()
+  )
+}
+
+export { toTitleCase }
